extract http headers
