+++ lib/rubyai/client.rb
@@ -1,10 +1,11 @@
 require 'faraday'
 require 'json'
-require_relative '../configuration'
+require_relative 'configuration'
+require_relative 'http'
 
 module RubyAI
   class Client
-    def initialize(api_key, messages, temperature: 0.7, model: Configuration::DEFAULT_MODEL)
+    def initialize(api_key, messages, temperature: 0.7, model: RubyAI::Configuration::DEFAULT_MODEL)
       @api_key = api_key
       @messages = messages
       @temperature = temperature
@@ -13,24 +14,15 @@ module RubyAI
 
     def call
       response = connection.post do |req|
-        req.url Configuration::BASE_URL
-        req.headers['Content-Type'] = 'application/json'
-        req.headers['Authorization'] = "Bearer #{@api_key}"
-        req.body = body.to_json
+        req.url RubyAI::Configuration::BASE_URL
+        req.headers.merge!(RubyAI::HTTP.build_headers(@api_key))
+        req.body = RubyAI::HTTP.build_body(@messages, @model, @temperature).to_json
       end
 
       JSON.parse(response.body)
     end
 
     private
-
-    def body
-      {
-        'model': Configuration::MODELS[@model],
-        'messages': [{"role": "user", "content": @messages}],
-        'temperature': @temperature
-      }
-    end
 
     def connection
       @connection ||= Faraday.new do |faraday|
